+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 * Feat: [Web] Add Twilio Device `updateToken(String)` function to allow updating of active device tokens.
 * Fix: [Web] Twilio Device does not unregister on `unregister()` method call due to 'device.off' not visible in js object causing device event listeners to remain attached on unregistered device.
 * Feat: [Web] Check unnecessary updating device with the same token. 
-* Feat: [Web] Update internal [CallStatus] mapping method. 
+* Feat: [Web] Update internal [CallStatus] mapping method.
+* Fix: [Web] Await Twilio Device `register()` and `unregister()` method calls.
 * Feat: update example.
 * Docs: update CHANGELOG